added Main, CardBlock, CardItems, Header, fixed bugs
