+++ app/page.js
@@ -1,4 +1,3 @@
-"use client"
 import Main from "@/components/Main/Main";
 
 export default function Home() {
