+++ components/Main/Main.jsx
@@ -1,11 +1,11 @@
 "use client"
 import React from 'react';
 import Header from "@/components/Header/Header";
-import CardItems from "@/components/CardItems/CardItems";
+import CardItems from "@/components/Main/CardItems/CardItems";
 
 const Main = () => {
     return (
-        <section style={{maxWidth: '320px', margin:'0 auto',  background: '#0F1B29' }}>
+        <section className='max-w-[320px] m-y-0 mx-auto bg-[#0f1b29]'>
             <Header />
             <CardItems />
         </section>
